Move testing files. Rename and move activity tab. Move password change metric component
