+++ bitwarden_license/bit-web/src/app/dirt/access-intelligence/v2/activity-tab/password-change-metric-v2/password-change-metric-v2.component.spec.ts
@@ -16,7 +16,7 @@ import { CipherId, OrganizationId, SecurityTaskId } from "@bitwarden/common/type
 import { SecurityTask, SecurityTaskStatus, SecurityTaskType } from "@bitwarden/common/vault/tasks";
 import { ToastService } from "@bitwarden/components";
 
-import { AccessSecurityTasksService } from "../services/abstractions/access-security-tasks.service";
+import { AccessSecurityTasksService } from "../../services/abstractions/access-security-tasks.service";
 
 import {
   PasswordChangeMetricV2Component,
