+++ bitwarden_license/bit-web/src/app/dirt/access-intelligence/v2/activity-tab/password-change-metric-v2/password-change-metric-v2.component.ts
@@ -25,7 +25,7 @@ import {
   TypographyModule,
 } from "@bitwarden/components";
 
-import { AccessSecurityTasksService } from "../services/abstractions/access-security-tasks.service";
+import { AccessSecurityTasksService } from "../../services/abstractions/access-security-tasks.service";
 
 export const PasswordChangeView = {
   EMPTY: "empty",
