+++ bitwarden_license/bit-web/src/app/dirt/access-intelligence/v2/new-applications-dialog-v2/new-applications-dialog-v2.component.stories.ts
@@ -15,7 +15,7 @@ import {
   MockSecurityTasksService,
   MockToastService,
   createAccessIntelligenceI18nMock,
-} from "../shared/test-helpers";
+} from "../testing";
 
 import {
   NewApplicationsDialogV2Component,
